refactor: make AssignmentLogger to not be a pydantic model
There is no reason for AssignmentLogger to be a pydantic model because
it's not a model at all (it's behavior, not data).

Remove inheritance from the base model for AssignmentLogger and allow
arbitrary types on Config instead. Also exclude assignment_logger from
pydantic serialization.

This refactor makes it easier to work with AssignmentLogger.
Especially when mocking it, as pydantic models are not easily mockable.