FF-3143 feat: add caching logger for assignment and bandit events (#68)
* refactor: remove unnecessary evaluation checks

BanditEvaluator.evaluate_bandit() always returns a result, so `is
None` checks were unnecessary. Furthermore, `evaluation` fields are
accessed without this guard a couple of lines below.

* refactor: make AssignmentLogger to not be a pydantic model

There is no reason for AssignmentLogger to be a pydantic model because
it's not a model at all (it's behavior, not data).

Remove inheritance from the base model for AssignmentLogger and allow
arbitrary types on Config instead. Also exclude assignment_logger from
pydantic serialization.

This refactor makes it easier to work with AssignmentLogger.
Especially when mocking it, as pydantic models are not easily mockable.

* feat: add assignment cache logger

* docs: add documentation for de-duplicating assignment and bandit actions

* chore: move cachetools from hard to test requirements

* chore: bump version

* refactor: spell out mk as make